start to fixup buildsystem
- check only for packages that are used
- fix addlibs file for cl
- fix include of Makemain.inc
- remove old addlibs.in file
- test builds only if OsiCbc is present
- fix compiler warnings
- remove unused defines from compiler flags
- remove unnecessary config_default.h

+++ FlopCpp/src/MP_model.cpp
@@ -263,9 +263,9 @@ void MP_model::attach(OsiSolverInterface *_solver) {
 	    (*i)->coefficients(f);
 	}
     }
-    nz = coefs.size();
+    nz = static_cast<int>(coefs.size());
 
-    messenger->statistics(Constraints.size(),m,Variables.size(),n,nz);
+    messenger->statistics(static_cast<int>(Constraints.size()),m,static_cast<int>(Variables.size()),n,nz);
 
     Elm = new double[nz]; 
     Rnr = new int[nz];
